Midi: strongly typed Midi::Note and Midi::Channel
the initial spark for the idea was me still finding places in the code
which still used hard coded digits for MIDI channels.

Refactoring the code with strongly typed notes and channels will make
our code base more readable and easy to maintain. It already unveiled
some minor bugs in real time MIDI note handling and fixed some dormant
issues due to missing range checks (e.g. when creating more than 128-36
instruments).

+++ src/core/IO/CoreMidiDriver.cpp
@@ -233,7 +233,7 @@ void CoreMidiDriver::sendNoteOnMessage( const MidiMessage& msg )
 
 	packetList.packet->timeStamp = 0;
 	packetList.packet->length = 3;
-	packetList.packet->data[0] = 0x90 | msg.getChannel();
+	packetList.packet->data[0] = 0x90 | static_cast<int>( msg.getChannel() );
 	packetList.packet->data[1] = msg.getData1();
 	packetList.packet->data[2] = msg.getData2();
 
@@ -252,7 +252,7 @@ void CoreMidiDriver::sendNoteOffMessage( const MidiMessage& msg )
 
 	packetList.packet->timeStamp = 0;
 	packetList.packet->length = 3;
-	packetList.packet->data[0] = 0x80 | msg.getChannel();
+	packetList.packet->data[0] = 0x80 | static_cast<int>( msg.getChannel() );
 	packetList.packet->data[1] = msg.getData1();
 	packetList.packet->data[2] = msg.getData2();
 
@@ -270,7 +270,7 @@ void CoreMidiDriver::sendControlChangeMessage( const MidiMessage& msg ) {
 
 	packetList.packet->timeStamp = 0;
 	packetList.packet->length = 3;
-	packetList.packet->data[0] = 0xB0 | msg.getChannel();
+	packetList.packet->data[0] = 0xB0 | static_cast<int>( msg.getChannel() );
 	packetList.packet->data[1] = msg.getData1();
 	packetList.packet->data[2] = msg.getData2();
