+++ src/core/IO/JackMidiDriver.cpp
@@ -120,7 +120,7 @@ void JackMidiDriver::JackMidiWrite( jack_nframes_t nframes ) {
 void JackMidiDriver::sendControlChangeMessage( const MidiMessage& msg ) {
 	uint8_t buffer[4];	
 	
-	buffer[0] = 0xB0 | msg.getChannel();
+	buffer[0] = 0xB0 | static_cast<int>( msg.getChannel() );
 	buffer[1] = msg.getData1();
 	buffer[2] = msg.getData2();
 	buffer[3] = 0;
@@ -342,7 +342,7 @@ void JackMidiDriver::sendNoteOnMessage( const MidiMessage& msg ) {
 
 	uint8_t buffer[4];
 
-	buffer[0] = 0x90 | msg.getChannel();	/* note on */
+	buffer[0] = 0x90 | static_cast<int>( msg.getChannel() );	/* note on */
 	buffer[1] = msg.getData1();
 	buffer[2] = msg.getData2();
 	buffer[3] = 0;
@@ -353,7 +353,7 @@ void JackMidiDriver::sendNoteOnMessage( const MidiMessage& msg ) {
 void JackMidiDriver::sendNoteOffMessage( const MidiMessage& msg ) {
 	uint8_t buffer[4];
 
-	buffer[0] = 0x80 | msg.getChannel();	/* note off */
+	buffer[0] = 0x80 | static_cast<int>( msg.getChannel() );	/* note off */
 	buffer[1] = msg.getData1();
 	buffer[2] = 0;
 	buffer[3] = 0;
